requirements.txt
joblib==1.3.2
numpy==1.26.1
opencv_python==4.8.1.78
pandas==2.1.2
pydantic==2.4.2
pytesseract==0.3.10
PyYAML==6.0.1
torch==2.1.0
torchvision==0.16.0

Pillow~=10.1.0
opencv-python~=4.8.1.78